feat: project detail tabs as nested App Router routes
- Move shell into projects/[id]/layout with path-derived active tab
- Add ProjectContext for read-only state and plan capability flags
- Add child routes: spaces, quotation, expenses, purchases, payments, documents, notes
- Root html: data-scroll-behavior for Next.js with smooth scroll CSS
- Budget thumbnails: next/image fill to fix aspect-ratio warning
